#315: Reorganize components like biology
