+++ components/molecules/GitHubIcon.vue
