+++ layouts/default.vue
@@ -40,19 +40,19 @@
 
 
 import LazyHydrate from 'vue-lazy-hydration';
-import BlogSection from '@/components/BlogSection.vue';
-import BottomDots from '@/components/BottomDots.vue';
-import Footer from './partials/Footer.vue';
-import Header from './partials/Header.vue';
+import BottomDots from '@/components/atoms/BottomDots.vue';
+import BlogSection from '@/components/organisms/BlogSection.vue';
+import Footer from '@/components/organisms/Footer.vue';
+import Header from '@/components/organisms/Header.vue';
 
 export default {
 
   components: {
+    BottomDots,
+    BlogSection,
     Footer,
     Header,
     LazyHydrate,
-    BottomDots,
-    BlogSection,
   },
 };
 </script>
