+++ pages/index.vue
@@ -19,9 +19,9 @@
 
 import LazyHydrate from 'vue-lazy-hydration';
 
-import Hero from '@/components/partials/Hero.vue';
-import PostImages from '@/components/partials/PostImages.vue';
-import Projects from '@/components/partials/Projects.vue';
+import Hero from '@/components/organisms/Hero.vue';
+import PostImages from '@/components/organisms/PostImages.vue';
+import Projects from '@/components/organisms/Projects.vue';
 
 export default {
   components: {
